+++ src/main/java/com/vedisoft/vedisoft2020/jparepository/StudentBatchJpaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.vedisoft.vedisoft2020.pojos.StudentBatch;
 
 @Repository
-public interface StudentBatchJpaRepository extends JpaRepository<StudentBatch, Integer> {
+public interface StudentBatchJpaRepository extends JpaRepository<StudentBatch, Long> {
 
 }
